move numlimbs to utils/ec
